+++ src/main/java/com/zjy/springbootadmin/entity/User.java
@@ -5,8 +5,12 @@ import com.baomidou.mybatisplus.annotation.TableId;
 import com.baomidou.mybatisplus.annotation.TableName;
 import java.io.Serializable;
 import java.time.LocalDateTime;
+
+import com.fasterxml.jackson.annotation.JsonIgnore;
 import io.swagger.annotations.ApiModel;
 import io.swagger.annotations.ApiModelProperty;
+import lombok.Getter;
+import lombok.Setter;
 
 /**
  * <p>
@@ -14,8 +18,10 @@ import io.swagger.annotations.ApiModelProperty;
  * </p>
  *
  * @author zjy
- * @since 2022-08-15
+ * @since 2022-08-16
  */
+@Getter
+@Setter
 @TableName("sys_user")
 @ApiModel(value = "User对象", description = "")
 public class User implements Serializable {
@@ -23,13 +29,14 @@ public class User implements Serializable {
     private static final long serialVersionUID = 1L;
 
       @ApiModelProperty("主键")
-        @TableId(value = "id", type = IdType.AUTO)
+      @TableId(value = "id", type = IdType.AUTO)
       private Integer id;
 
       @ApiModelProperty("用户名")
       private String username;
 
-    private String password;
+      @JsonIgnore
+      private String password;
 
       @ApiModelProperty("昵称")
       private String nickname;
@@ -46,82 +53,5 @@ public class User implements Serializable {
       @ApiModelProperty("创建时间")
       private LocalDateTime createTime;
 
-    
-    public Integer getId() {
-        return id;
-    }
 
-      public void setId(Integer id) {
-          this.id = id;
-      }
-    
-    public String getUsername() {
-        return username;
-    }
-
-      public void setUsername(String username) {
-          this.username = username;
-      }
-    
-    public String getPassword() {
-        return password;
-    }
-
-      public void setPassword(String password) {
-          this.password = password;
-      }
-    
-    public String getNickname() {
-        return nickname;
-    }
-
-      public void setNickname(String nickname) {
-          this.nickname = nickname;
-      }
-    
-    public String getEmail() {
-        return email;
-    }
-
-      public void setEmail(String email) {
-          this.email = email;
-      }
-    
-    public String getPhone() {
-        return phone;
-    }
-
-      public void setPhone(String phone) {
-          this.phone = phone;
-      }
-    
-    public String getAddress() {
-        return address;
-    }
-
-      public void setAddress(String address) {
-          this.address = address;
-      }
-    
-    public LocalDateTime getCreateTime() {
-        return createTime;
-    }
-
-      public void setCreateTime(LocalDateTime createTime) {
-          this.createTime = createTime;
-      }
-
-    @Override
-    public String toString() {
-        return "User{" +
-              "id=" + id +
-                  ", username=" + username +
-                  ", password=" + password +
-                  ", nickname=" + nickname +
-                  ", email=" + email +
-                  ", phone=" + phone +
-                  ", address=" + address +
-                  ", createTime=" + createTime +
-              "}";
-    }
 }
