+++ src/main/java/com/zjy/springbootadmin/utils/generator.java
@@ -11,7 +11,7 @@ import java.util.Collections;
  * by halfling
  * @since 2022-8-12
  */
-public class generater {
+public class generator {
     public static void main(String[] args) {
         generate();
     }
@@ -29,6 +29,10 @@ public class generater {
                             .pathInfo(Collections.singletonMap(OutputFile.mapper, "D:\\projects\\SpringBoot\\springbootAdmin\\src\\main\\resources\\mapper")); // 设置mapperXml生成路径
                 })
                 .strategyConfig(builder -> {
+                    builder.entityBuilder().enableLombok();     //lombok
+                    builder.mapperBuilder().enableMapperAnnotation().build();   //mapper
+                    builder.controllerBuilder().enableHyphenStyle()             //驼峰
+                            .enableRestStyle();                                 //RestController
                     builder.addInclude("sys_user") // 设置需要生成的表名
                             .addTablePrefix("t_", "sys_"); // 设置过滤表前缀
                 })
