+++ src/main/resources/application.yml
@@ -1,5 +1,5 @@
 server:
-  port: 9999
+  port: 9090
 spring:
   datasource:
     driver-class-name: com.mysql.cj.jdbc.Driver
