+++ vue/package.json
@@ -11,7 +11,8 @@
     "core-js": "^3.8.3",
     "element-ui": "^2.15.9",
     "vue": "^2.6.14",
-    "vue-router": "^3.5.1"
+    "vue-router": "^3.5.1",
+    "vuex": "^3.6.2"
   },
   "devDependencies": {
     "@vue/cli-plugin-babel": "~5.0.0",
