+++ vue/src/main.js
@@ -1,6 +1,7 @@
 import Vue from 'vue'
 import App from './App.vue'
 import router from './router'
+import store from './store'
 import ElementUI from 'element-ui';
 import 'element-ui/lib/theme-chalk/index.css';
 import './assets/globle.css'
@@ -14,5 +15,6 @@ Vue.prototype.request=request  //axios的request组件
 
 new Vue({
   router,
+  store,      //引入vuex
   render: h => h(App)
 }).$mount('#app')
