+++ vue/src/router/index.js
@@ -1,15 +1,24 @@
 import Vue from 'vue'
+import store from "@/store";
 import VueRouter from 'vue-router'
-import HomeView from '../views/HomeView.vue'
+import HomeView from '../views/Manage.vue'
+import Home from "@/views/Home";
 
 Vue.use(VueRouter)
 
 const routes = [
   {
     path: '/',
-    name: 'home',
-    component: HomeView
+    name: 'Manage',
+    component: () => import('../views/Manage'),
+    redirect: '/Home',  //默认重定向
+    // 路由
+    children: [
+      {path: 'User', name: '主页', component: () => import('../views/User'),},
+      {path: 'Home', name: '主页', component: () => import('../views/Home')}
+    ]
   },
+
   {
     path: '/about',
     name: 'about',
@@ -22,6 +31,13 @@ const routes = [
 
 const router = new VueRouter({
   routes
+})
+
+//vuex所需路由守卫
+router.beforeEach((to, from, next) => {
+  localStorage.setItem("currentPathName", to.name)    //设置当前路由名称, 为了在Header组件中使用
+  store.commit("setPath")                             //触发store的数据更新
+  next()                                              //放行路由
 })
 
 export default router
