+++ vue/src/views/Home.vue
@@ -1,0 +1,15 @@
+<template>
+  <div>
+    <h1>主页</h1>
+  </div>
+</template>
+
+<script>
+export default {
+  name: "Home"
+}
+</script>
+
+<style scoped>
+
+</style>
